Refactoring into separate files
