+++ .travis.yml
@@ -13,7 +13,7 @@ before_script:
 
 after_success:
 - bash scripts/build.sh
-- python3 scripts/command_coverage.py
+- python3 scripts/command_coverage/runner.py
 
 after_failure:
 - cat test_result | python scripts/send_to_bot.py
